+++ src/common/types/common.ts
@@ -1,7 +1,7 @@
 export type Nullable<T> = T | null;
 export type Percentage<T extends number | string> = `${T}%`;
 export type Percentable<T extends number | string> = T | Percentage<T>;
-export type Functionable<T> = T | (() => T);
+export type Functionable<T, TArgs extends any[] = any[]> = T | ((...args: TArgs) => T);
 export type Arrayable<T> = T | T[];
 export type Autocomplete<T extends string> = T | (string & {});