+++ src/components/common/media/icons/Badge/index.ts
@@ -1,0 +1,2 @@
+import './Badge.scss';
+export { default } from './Badge';