+++ src/components/domain/ModelCard/index.ts
@@ -1,3 +1,3 @@
 import './ModelCard.scss';
 export { default } from './ModelCard';
-export type { ModelCardProps, BaseModelCardProps } from './ModelCard';
+export type { ModelCardProps, BaseModelCardProps, ModelRarityTabGroup } from './types';