+++ src/services/SearchService/base/types.ts
@@ -1,0 +1,11 @@
+export type ScoreColorProps = {
+  scores: [
+    min: number,
+    current: number,
+    max: number,
+  ],
+  colors: [
+    low: string,
+    high: string,
+  ]
+}