+++ src/services/SearchService/index.ts
@@ -1,2 +1,2 @@
-export { SearchResult, SearchResultItem, LastResult } from './types';
-export * from './SearchService';
+export * from './artifact';
+export * from './weapon';