+++ change-log.md
@@ -1,10 +1,15 @@
-# 🎉 Danho's Genshin Tool v . 2.1.14 🎉
+# 🎉 Danho's Genshin Tool v . 2.2.0 🎉
+
+## Reworked Recommended Weapons
+* Moves signature weapon into the tabbar
+* More accurately recommends weapons based on internal calculation score (functions like Artifact Helper)
 
 ## Updates
 * Added "Off-field Damage" tag to Columbina
 
 ## Fixes
 * Recommended weapons not consistently rendering tabs as rarity hierarchy
+* Wolf-Fang being marked as "Rare" instead of "Epic"
 * Nod-Krai filter not being in SearchableMobList
 * Sigewinne's bonus talents not showing off-field damage
 * Toukabou Shigure marked as Legendary, when it should be Epic